+++ flask/cli.py
@@ -14,6 +14,7 @@ from __future__ import print_function
 import ast
 import inspect
 import os
+import platform
 import re
 import ssl
 import sys
@@ -260,10 +261,16 @@ def locate_app(script_info, module_name, app_name, raise_if_not_found=True):
 def get_version(ctx, param, value):
     if not value or ctx.resilient_parsing:
         return
-    message = 'Flask %(version)s\nPython %(python_version)s'
+    import werkzeug
+    message = (
+        'Python %(python)s\n'
+        'Flask %(flask)s\n'
+        'Werkzeug %(werkzeug)s'
+    )
     click.echo(message % {
-        'version': __version__,
-        'python_version': sys.version,
+        'python': platform.python_version(),
+        'flask': __version__,
+        'werkzeug': werkzeug.__version__,
     }, color=ctx.color)
     ctx.exit()
 
@@ -333,7 +340,7 @@ class DispatchingApp(object):
 
 
 class ScriptInfo(object):
-    """Help object to deal with Flask applications.  This is usually not
+    """Helper object to deal with Flask applications.  This is usually not
     necessary to interface with as it's used internally in the dispatching
     to click.  In future versions of Flask this object will most likely play
     a bigger role.  Typically it's created automatically by the
@@ -370,7 +377,7 @@ class ScriptInfo(object):
             app = call_factory(self, self.create_app)
         else:
             if self.app_import_path:
-                path, name = (self.app_import_path.split(':', 1) + [None])[:2]
+                path, name = (re.split(r':(?![\\/])', self.app_import_path, 1) + [None])[:2]
                 import_name = prepare_import(path)
                 app = locate_app(self, import_name, name)
             else:
@@ -725,7 +732,7 @@ def _validate_key(ctx, param, value):
     return value
 
 
-@click.command('run', short_help='Runs a development server.')
+@click.command('run', short_help='Run a development server.')
 @click.option('--host', '-h', default='127.0.0.1',
               help='The interface to bind to.')
 @click.option('--port', '-p', default=5000,
@@ -777,10 +784,10 @@ def run_command(info, host, port, reload, debugger, eager_loading,
                threaded=with_threads, ssl_context=cert)
 
 
-@click.command('shell', short_help='Runs a shell in the app context.')
+@click.command('shell', short_help='Run a shell in the app context.')
 @with_appcontext
 def shell_command():
-    """Runs an interactive Python shell in the context of a given
+    """Run an interactive Python shell in the context of a given
     Flask application.  The application will populate the default
     namespace of this shell according to it's configuration.